rawexec: SIGINT is not available on Windows
